ENYO-2132: Fix moon.Panels focus lost on right key press
Issue:
When pressing right button on panel and there is no control to get
focused. Focus disappears and current is set to panel.

Fix:
Checking next panel on spotlightRight handler is removed from previous
moon.Panels. We revert the code so it is not return true when there is
no next panel.

Signed-off-by: Kunmyon Choi <kunmyon.choi@lge.com>