ENYO-2122: Refine fix to support setIndex() at creation time
For 2.6.0-pre.15, we merged a fix that ensured proper panel layout
in the case where a developer had set a non-zero index at creation
time (either declaratively or imperatively).

This improved fix accomplishes the same goal with logic that is
more specifically targeted to the nature of the original issue and
should be less brittle.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)