The Video component would not always successfully start playing the video after a rewind operation.
This problem was, at least in part, caused by the need to emulate rewind
on platforms that don't support negative `playbackRate`s.  Additionally,
a problem was discovered with handling the switch between slow and
fastForward. The slowForward playback rate array was updated to remove
the rate of '1' (normal speed) and the fastforward mechanism is updated
to move to play when reaching the end of the playback rate array
(Previously, it rotated to the fast forward array after playing the '1'
rate). There should be no change for apps that did not alter the
playback rate array. The player has also been updated so that rewinding
after a slowForward will now result in a slowRewind. The converse also
applies.