ENYO-3356 Panels: getActive() is not existed when just one panel is replaced
Issue

In specific user case of Device Connector application, if
'setLastFocusChild' is called when value of getActive() of panels is
undefined, error is occurred.
At that time, next panel we want to replace is not displayed.

Fix

It seems to need to check if active panel is existed or not before
spotlight moves to last control.
Plus, Spotlight moves this.getActive() after replacing panel.

ENYO-3356 Panels: getActive() is not existed when just one panel is
replaced
Enyo-DCO-1.1-Signed-off-by: Sangwook Lee sangwook1203.lee@lge.com