ENYO-3622: moon.Panels screen broken while transition by key press on breadcrumb with rtl and accessibility on
Condition:
Rtl, accessibility on

Issue:
We change left position of the first panel and set overflow visible when
move back to first panel to hide breadcrumb area without any size
changes for each panel.
If move forward to next panel after press 5way enter key on breadcrumb
to move back to previous panel, panel position shifted left/right or
first panel is showing instead of second.
By intensive testing, we noticed that this happens when breadcrumbs is
set to display none with DOM.focus on one of the breadcrumb. And also
noticed that this issue is not happens if we are not doing DOM.focus on
item of first panel when pressing enter to move next panel by set
accessibilityDisabled on item.

Fix:
Delay breadcrumbs hide on indexChange from before start transition to
after complete transition to avoid focused element hidden and make it
visible and scroll element in the screen.
(Not 100% sure about the root cause for now. But, below like is a hint.)
https://www.sitepoint.com/community/t/prevent-page-scrolling-down-on-form-focus-event/5341/2

Enyo-DCO-1.1-Signed-off-by: Kunmyon Choi (kunmyon.choi@lge.com)